Implement move semantics with .ref / .dispose()
This is how we will free memory used by arrays. Users are required to properly use or dispose of all arguments to functions, and to call `.ref` as needed.

This is a bit scuffed. It took me forever to work out all the manual reference-counting especially as it related to some of the trickier functional programming transforms. But hopefully it almost works now!

All tests are failing except those that exercise `grad()` and `linearize()`, i.e. those in `test/numpy.test.ts` and `test/tracing.test.ts`. Gotta work those out.

This took forever, but we're almost there? :D